Change protocol version to add size-based "headers" limit.
Introduce a new network-level limit on the size of "headers" messages.
In the current protocol inherited from Bitcoin, the maximum number of
headers in a single message is limited to 2,000.  This number is also
used to decide when to ask a peer for more.

Due to merge-mining, however, a message with 2,000 headers can be huge
(and may even be too large for the P2P message limit of 32 MiB).  This
patch introduces a new hard limit at 6 MiB for headers messages,
used for peers advertising the new protocol version 110000.

In addition, it introduces a soft "threshold limit" of 4 MiB.  The
client sends at least that many headers if it has more, so this is used
as additional criterion to decide when to ask for more (if the peer has
the new protocol version set).

See also https://github.com/namecoin/namecoin-core/pull/32.