Do not enforce IsInitialBlockDownload() check for testnet.
The (changed?) logic is that we consider everything initial block
download if the last known block is more than a day old.  This can,
however, happen in a legitimate situation on testnet if noone is mining.
Even more, this prevents future restarting of mining.  Thus disable this
check for testnet.