Introduce an earlier check for invalid Namecoin tx.
This fixes #116.  The crash described there happens when adding an
invalid transaction to the mempool, since the constructor of
CTxMempoolEntry assumes that we have a valid structure for name tx (at
most one name input/output).  This is checked only later at the moment,
failing an assertion.

This patch introduces an earlier check for the validity of a Namecoin
transaction when adding to mempool, so that this assumption is actually
valid later on.  The check still happens as part of CheckInputs, which
is redundant for the mempool addition (not block verification) but that
is no problem.