Fix name_update for the case of very low wallet balance.
The wallet code contains a check that previously failed with
"insufficient balance" when a name_update was performed with the wallet
balance less than what is locked in a name (0.01 NMC).  The check did
not count the name input towards the available balance.

This patch fixes that, so that name_update works also with very low
wallet balance of <0.01 NMC.  It also adds a regtest for this case.