Fix Makefile.am's txcreatesign and blanktx dependencies.
This commit needs to be reverted after AAA is deployed.

This will hopefully fix the Travis CI failures on the dev branch.

Quoting @domob1812: "Some time ago Bitcoin introduced tests for `bitcoin-tx` that change the version number of the created transactions.  Since a change here is incompatible to Namecoin (because Namecoin uses a fixed `nVersion` to distinguish Namecoin transactions; something I proposed to remove with a fork some time ago in the forum), I've "reverted" these changes.  Thus this file [txcreatesignv1.hex] does not exist for Namecoin, it is instead `test/data/txcreatesign.hex`."

refs #142