Add more testing for mempool cleanup.
Add more testing for the proper clean up of the mempool after name
conflicts.  The existing tests only covered cases when the conflicted tx
was actually mined and reorged, the new test is for the situation that
it was always just in the mempool.

With the latest upstream merge, we fail to mark tx in the wallet as
conflicted more often, also update the tests for this.

See https://github.com/namecoin/namecoin-core/issues/175.