Fix bool values in JSON RPC results.
The change from push_back(Pair) to pushKV with UniValue's broke the
output of bool values in the JSON RPC interface, in particular "expired"
and "ismine".  Those are not supported correctly by pushKV, and were
returned as int's in JSON.

This commit reverts back to push_back(Pair) for these two keys, which
properly supports bool.  It also adds regtests verifying that these
fields are actually boolean.