Remove nameop support in createrawtransaction.
This commit removes the nameop argument that was present for
createrawtransaction, so that the RPC is now again a complete clone of
Bitcoin's functionality.

Instead, we will add Namecoin rawtx support with a new
namerawtransaction command, as suggested in
https://github.com/namecoin/namecoin-core/issues/181.

The regtests are not yet updated, so will currently fail with just this
commit (as some of them use createrawtransaction with nameop).  Also the
helper method AddRawTxNameOperation is still present in the code
(although now unused), as it will be used as basis for namerawtransaction.