Add namerawtransaction RPC command.
This new command accepts an already created raw transaction with inputs
and outputs, and modifies a given output to be a name operation.  This
replaces the functionality previously implemented directly in
createrawtransaction, but gives more flexibility for future changes.

See https://github.com/namecoin/namecoin-core/issues/181 for the
original proposal of this change.

The command for now only supports name_update operations (like
createrawtransaction used to), but will be extended with name_new and
name_update in the future as well.