Allow name_new raw tx to generate rand automatically.
This commit makes the "rand" option for creating name_new's with
namerawtransaction optional.  If not explicitly provided, the rand will
be generated randomly by the client.

The return value of namerawtransaction is changed from just the hex
string to a JSON object, that will hold the hex string in 'hex'.  For
name_new's, we additionally output the rand value in 'rand', so it can
be used by name_firstupdate when auto-generated.