Add src/rpc/names.cpp to check-rpc-mappings.py source list.
The script contrib/devtools/check-rpc-mappings.py verifies consistency
between RPC dispatch tables and the argument conversions listedn in
src/rpc/client.cpp.  For this, it needs a list of sources that contain
dispatch tables.  src/rpc/names.cpp was missing from this list, leading
to errors as in:

https://github.com/namecoin/namecoin-core/issues/203

This commit adds the missing source file, which makes the script run
successfully.  Note that src/wallet/rpcnames.cpp, which also contains
Namecoin RPCs, need not be listed; the dispatch table for those RPCs is
included in src/wallet/rpcwallet.cpp.