Remove the dblock limit inherited from Namecoin.
Since Namecoin was historically based on BDB, it has an artificial limit
on the number of db locks that can be used in each block (so that
LevelDB and old BDB clients are compatible).

This is no longer relevant for Chimaera, since the blockchain starts
fresh and uses only modern clients.  Thus we can remove the limit.

This solves https://github.com/chimaera/proto-chi/issues/12.