Change PoW algo to neoscrypt.
Add a new GetPowHash() function to CPureBlockHeader, which computes the
neoscrypt hash instead of SHA256D.  Use this for PoW verification (and
for mining itself in the generate command).

This does not yet update the regtests, and also does not update the
genesis blocks in chainparams yet (so they won't be valid for now).