Update block rewards and halving schedule.
As discussed in https://github.com/chimaera/proto-chi/issues/6, we want
to have one block per minute, halving every 4 years (so then every 2.1M
blocks) and an initial block subsidy of 10 CHI.

This also sets the MAX_MONEY constant to 560M, which is above the
555,555,556 coins estimated in the issue (including the premine later
on).

On regtest net, we keep the initial subsidy at 50 CHI, so that we don't
have to update a lot of constants in the tests.  To support this, we add
a new consensus parameter initialSubsidy instead of hardcoding the value
in validation.cpp.

Also, remove the nPowTargetTimespan param, as it is no longer needed (we
have continuous difficulty updating).