Add premine to P2SH to genesis block.
This reworks the genesis-block construction in chainparams.cpp, so that
the genesis block coinbase pays the premine of 333,333,333 CHI to a P2SH
multisig address.

This is the maximum possible supply, and might be reduced after the
presale by burning coins.  For more details, see:

  https://github.com/chimaera/proto-chi/issues/6

The P2SH address is currently set both on mainnet and testnet to a
1-of-2 address, with the privkeys given in the source.  This will be the
final address for testnet and regtest, used for testing.  The mainnet
address will be updated before launch, which is tracked here:

  https://github.com/chimaera/proto-chi/issues/14