Remove special "Namecoin" tx version.
Namecoin uses tx with a special nVersion set for name operations,
presumably due to legacy reasons.  This conflicts with upstream
Bitcoin's usage of nVersion=2 for opt-in RBF.  It is also totally
unnecessary, as one can get all information about name operations (or
not) from just looking at the output scripts of a tx.  Thus, this
"Namecoin tx version" concept is at most a performance optimisation for
some things (e. g., name_list or name_pending).

Since we start fresh with Chimaera, we can just remove it to make the
code closer to upstream Bitcoin.  This commit does this, which implements
https://github.com/chimaera/proto-chi/issues/19.