Set mainnet and testnet block reward to 1 CHI.
The reduced block reward is intended to be active during the presale
only.  After the total amount of sold coins is determined (which in turn
fixes the final supply), we will perform a hard fork that sets the
reward to a higher value that will yield the correct number of PoW
minable coins in the final supply.

This fixes https://github.com/chimaera/chimaera/issues/24.