Update name and value size limits.
This updates the maximum name length to 256 bytes (from 255) and the
maximum value size to 2048 bytes (from 1023).  It also increases the
maximum script element size, which is an effective upper bound for both,
to 2048 and adds static_asserts that verify that the script element size
is large enough for the two other limits.

This implements https://github.com/chimaera/chimaera/issues/23.