Set block time to 30s.
Update the block time to be 30s rather than 1m, and increase the halving
interval accordingly so that halvings still take place every 4 years.

Note that this increases (doubles) the coin supply excluding premine.
But since the initial block reward is very small anyway and will be
adjusted according to the outcome of the presale later, this does not
matter.