Switch activation of segwit and CSV from version bits.
Instead of using version bits to activate segwit and CSV, use the
BIP16Height to activate them all together with P2SH.  The height is set
to 0 on mainnet and testnet, but that will be updated for Namecoin.  The
height is set to 432 on regtest net, which immitates when BIP9
activation would hit so that the regtests work.

Some tests needed to be updated if they assumed that segwit is activated
right away by mining 500 blocks before running the test.  The segwit
regtests can be activated again with this, after removing logic that
tests the BIP9 stages (which are not present).

This allows to activate segwit and CSV with a scheduled soft fork, even
before AAA is launched.