Deflake chimaera_premine.py test.
getaddressesbylabel seems to return the addresses in a random order.
This leads to flaky failures of chimaera_premine.py, when the returned
order is not the one needed to generate the real premine address.

To fix this, we add the set of pubkeys used for the premine (with the
right order) and use that to generate the premine address.  We also
verify that the *set* (ignoring the order) of pubkeys matches the
hardcoded one.