Add explicit function to verify names / values.
Namecoin (and thus also Chimaera) enforces maximum name and value
lengths.  These were checked directly inline in the validation code and
RPC methods.

This commit introduces explicit IsNameValid / IsValueValid functions
that implement the checks and are called whereever needed from the main
code.

This way, we can explicitly test the functions in the unit tests, and
can more easily extend them in the future to implement more checks as per
https://github.com/chimaera/chimaera/issues/31.