Set REJECT_INVALID code and print error for invalid names/values.
When we reject a name transaction due to the name or value being
invalid, set the proper REJECT_INVALID code on the CValidationState
object.  Also print log errors that detail what failed.

This is a small follow-up for https://github.com/chimaera/chimaera/issues/31
to make the errors better debugable.