Fix reference to namecoin.conf in regtest.
This fixes a small oversight, where one particular regtest function
still referenced "namecoin.conf" instead of "chimaera.conf".  This led
to both files being created in the regtest node, but otherwise the tests
worked (presumably because the bind= option added this way to
namecoin.conf was not critical).

This is slightly related to https://github.com/chimaera/chimaera/issues/30,
not the fix (as it only affects the regtest framework).