Use getNameInfo for name_pending.
This refactors getNameInfo, such that the expiration logic is removed to
a separate function.  This allows to use getNameInfo also for
name_pending, which does not have expiration / height information.

By doing so, we simplify the code, make it more consistent and also make
it easier to support custom encoded names/values in the future for
https://github.com/namecoin/namecoin-core/issues/194.