Refactor help-text building for name info RPCs.
This replaces getNameInfoHelp() with a builder class and a fluent
interface.  This not only makes the code cleaner, but also allows a more
flexible interface so that we can now also use this for the help text of
name_pending, as well as properly document the "transferred" field of
name_list (https://github.com/namecoin/namecoin-core/issues/193).