Move getNamePrevout close to name_firstupdate.
getNamePrevout is a helper method that is only used by name_firstupdate,
so it makes sense to move it close to that place.  The "general front"
of rpcnames.cpp should be used for things that are used by multiple
methods, like the DestinationAddressHelper and more things in the
future.

This is a pure code move, together with also moving
MAX_NAME_PREVOUT_TRIALS into the function.