Encapsulate merge-mining RPC logic in AuxpowMiner.
This creates the AuxpowMiner class in a new pair of source files, which
encapsulates the previously-global state as well as the main logic
behind the merge-mining RPCs.

By doing so, this logic is nicer separated from the existing code.
Also, by having all the state in a class that is explicitly constructed
and destructed, we don't have a mess with global variables and static
variables in the function itself.  And finally, by having the code
separated, we can more easily add unit tests (at least for the block
regeneration logic, which is somewhat non-trivial) in a follow-up.

This is mostly a pure refactor and move of code.  It does, however,
remove the usage of "submitblock_StateCatcher" when submitting the
block; the caught validation state was not actually used anyway.

See https://github.com/namecoin/namecoin-core/issues/224.