Make name-script-to-JSON publicly accessible.
This moves the logic to convert a CNameScript to JSON from inside
ScriptPubKeyToUniv to its own, publicly accessible function.  By doing
so, we separate the name-related code a bit better off the core Bitcoin
code; also, we make it available for use in other places as well -- like
in listunspent for https://github.com/namecoin/namecoin-core/issues/192.