Deflake the recently-introduced tests for AuxpowMiner.
The unit tests for AuxpowMiner that were recently introduced as part of
https://github.com/namecoin/namecoin-core/issues/224 were flaky:

When a new best tip is found, all cached blocks are cleared.  This also
means that it is possible for a *newly constructed* block to actually
have the same pointer address.  Hence, the comparison for inequality by
address was incorrect; instead, we have to compare by block hash only in
this case.