Construct auxpow object before setting it.
This is a small and trivial refactoring:  Instead of setting the auxpow
object in a CBlockHeader and then initialising it in place, we construct
the full CAuxPow and only then set it in the block header as a final
step.  This decouples the construction code as much as possible from the
block header.