Create and manage a global AuxpowMiner instance.
This adds a global variable g_auxpow_miner for the singleton AuxpowMiner.
During the startup- and shutdown sequences this AuxpowMiner is constructed
and destructed, respectively.  (With the same code that is used in
upstream Namecoin, for future maintenance.)

This builds on the previous commit, which introduced the basic auxpow
code back into Chimaera.  It prepares for using the AuxpowMiner for
implementing getwork (https://github.com/xyonplatform/chimaera/issues/33)
in a follow-up.  The current behaviour of the daemon is still not changed
by this commit alone.