Add back getwork for mining.
This adds back an implementation of getwork (together with creatework
and submitwork, performing each of the two "forms" of getwork and without
the need for a local wallet).

This was removed in upstream Bitcoin in
https://github.com/bitcoin/bitcoin/pull/4100.  The new version, however,
is based on Namecoin's merge-mining code and thus easy to maintain.
It is also necessary to support stand-alone mining with PoW data in the
future out-of-the-box with existing mining software.

I've tested that it works on testnet with our modified cpuminer.  Proper
regtests for it will be added as a follow-up.

Implements https://github.com/xyonplatform/chimaera/issues/33, except
for the still missing regtests.