Refactor ismine computation for name_pending.
This is a pure refactor of the way in which name_pending computes the
ismine field.  It moves the main logic to new helper methods, and
introduces in particular a helper class that wraps the wallet *if
available* for the non-wallet RPCs (where it provides optional extra
functionality).

Based on this, we can in a next step use the logic for other RPCs as
well (in particular, name_list to resolve
https://github.com/namecoin/namecoin-core/issues/219).