Change name_list's transferred to ismine.
name_list's response has an output field 'transferred', which indicates
whether the name is *not* owned by the wallet (anymore).  name_pending
has a similar field, but that is called 'ismine' and it indicates
whether or not the name *is* owned by the wallet.

This commit changes name_list's field to be 'ismine' as well, with the
corresponding negation of the bool.  This makes the fields consistent,
and furthermore avoids the confusing "inverse logic" involved in the old
'transferred' field.

This resolves https://github.com/namecoin/namecoin-core/issues/219.