Correctly handle unconfirmed name_new in mining.
When creating blocks from the mempool, miners need to make sure that they
do not include name_firstupdate transactions with still immature name_new
inputs; those are allowed in the mempool but not in the blockchain (but
they will be fine once the name_new is mature enough).

There is already logic to ensure that, but it had a bug that did not
prevent name_firstupdate's to be included in blocks if the corresponding
name_new was actually *unconfirmed*.

This caused https://github.com/namecoin/namecoin-core/issues/50 and
presumably also https://github.com/namecoin/namecoin-core/issues/227.

With this commit, we fix the underlying bug and add regression tests
that exercise and cover this scenario (as well as general handling of
immature / unconfirmed inputs to name operations).