Remove pblock argument to GetNextWorkRequired.
The difficulty retargeting algorithm we use in Xyon (DGW) does not need
to know the current block.  As such, the pblock argument to
GetNextWorkRequired was already unused.

This removes it completely, which more explicitly enforces that it is
not used at all (and this assumption will not be broken in the future
accidentally).  This will make it easier to support dual-algo mining in
the future, where the difficulty depends on the chosen algorithm.