Read block headers from disk when needed.
Instead of constructing the block headers from CBlockIndex instances
out of in-memory data, read the block headers from disk instead.  This
is needed in the future when we have PoW data attached to the headers,
which is not kept in memory.

This reverts one piece of b841843b3c506be536659b96c35b132c09088cab
(which removed auxpow support inherited from Namecoin).