Use nTime for CPureBlockHeader::IsNull.
CPureBlockHeader::IsNull looks at nBits as a proxy for whether a block
is "null" (rather than checking all fields).  This no longer works with
the proposed mining system in Xyon, as we move nBits from the actual
block header to the PoW data.  The nBits in the block header itself are
always zero, and thus this method would always consider blocks as null.

Instead of using nBits, we can just as well use nTime.  That is also
guaranteed to not be zero for a valid block, and is just as efficient to
check (as well as not impacted by any mining changes).