Add source files for PowData class.
This adds a new pair of source files powdata.h/.cpp, which hold the new
PowData class as well as some additional stuff (like the PowAlgo enum).
This implements the core logic we will need for PoW data, including the
serialisation / deserialisation as specified in our mining plan and the
logic to validate a PoW data object.

This commit also adds basic unit tests for the new code, but does not yet
integrate it into the codebase otherwise.  In particular, the PowData is
not yet used anywhere, and the consensus protocol is not yet changed.