Use PowData for block PoW.
This is the main change that switches block validation over to PoW data
(as described in our mining design and required for
https://github.com/xyonplatform/xyon/issues/36).  We change CBlockHeader
to be not just a CPureBlockHeader, but hold a PowData instance in addition.

The nBits of blocks themselves are set to zero, and the difficulty is
instead stored and read from the PowData's nBits.  Mining is done through
the nNonce of the PoW data's "fake header", and validation of a block's
PoW is done through the attached PoW data.

This implements the main change and makes the unit tests pass, but is
otherwise not complete.  The regtests still need to be fixed in a
follow-up change, and also the RPC interface needs to be updated (a new
'powdata' field for getblock needs to be added, and the getwork-type RPCs
for mining need to be adapted as well).