Remove nNonce of CBlockIndex.
The CBlockIndex instances are meant to hold the tree of "block headers"
in memory.  But with the switch to PoW data, block headers have always
a zero nonce; so it is not necessary to store that in the CBlockIndex.

The block hash can still be computed, and the difficulty is also kept
(to determine total chain work).  The PoW, on the other hand, cannot be
verified anyway from the CBlockIndex, as it misses the PowData (which is
not kept in memory).