Add powdata field for getblock RPC.
We now return a JSON-version of the PoW data of a block in the new
'powdata' field of the 'getblock' RPC method.  Note that getblockheader
does not return the powdata, as it is based off the in-memory
CBlockIndex graph; we would have to change that to read blocks from disk
to get the PoW data, which seems unnecessary for now (the same is true
for the 'auxpow' field in Namecoin).