Update getwork-type RPCs to work with PoW data.
After introducing PoW data in previous changes, this adapts the
getwork-type mining RPCs to work with the new structure.  This includes
two main parts:

First, creatework is changed to construct a PoW data instance with a
fake header instead of returning the actual block header as 'data'.
submitwork is similarly changed to parse the solved data as fake header
and set it on the block.

Second, since now the block hash does not change through mining, we key
the cached blocks by it similar to how getauxblock works.  This adds the
block hash to the result of creatework, and expects the hash as a new
first argument for submitwork.