Support PoW data in the mininode test framework.
With the introduction of PoW data in the core codebase, the
serialisation format for block headers was changed, and the difficulty
of blocks is now carried by nBits on the PoW data, not the header itself.

This implements the corresponding change also in the mininode messages and
the regtests themselves.