Update REST test to read only pure block header.
With the introduction of PoW data, the "full" block headers are no
longer 80 bytes in size (larger now instead).  This updates the interface_rest
regtest to take that into account.

It basically reverts the corresponding part of
b841843b3c506be536659b96c35b132c09088cab, and brings the test back in
line with what upstream Namecoin has because of auxpow.