Add algo argument to GetPowHash().
Instead of always computing the Neoscrypt hash in GetPowHash(), this adds
a PowAlgo argument to specify the algorithm that should be used.  Both
SHA256D and Neoscrypt are supported by the function.

All places that call it have been updated accordingly; most of them are
simply passing in the core PoW algo from the block's PoW data.  (So that
the code will work seamlessly with real dual-algo by just setting the
algo on the PoW data correctly.)

Note that this does not yet change the consensus to allow both algos; it
just refactors the core code to support that change in the future.