Make verification of nBits optional.
In the same manner as fCheckPow allows turning off PoW validation for
TestBlockValidity (which is used to verify blocks constructed but not
yet mined), we add a new flag fCheckBits that can be used to turn off
validation of nBits against the expected difficulty.  This is required
to test a block's validity before the mining algorithm (and thus
difficulty) has been fixed in the future.