Force explicit specification of algo in PowData.
Previously, we already set the mining algo to NEOSCRYPT in the
constructor of PowData, and there was no way to change that.  In this
change, we initialise the algorithm to INVALID and provide a setCoreAlgo
method.  This forces all places in the code that construct a PowData
object to explicitly specify the algorithm (as NEOSCRYPT for now).

To support this through CBlockTemplate, we compute expected difficulties
for each mining algorithm and store it in the CBlockTemplate.  This
allows us to provide a SelectAlgo method on CBlockTemplate that can
update just the mining algo and difficulty for the constructed block
after the fact.