Return bits for all algos from getblocktemplate.
Since the difficulty is different for each mining algo (or will be soon
when we switch to real dual-algo), we have to return all difficulties
from getblocktemplate.  To this end, the 'bits' field is changed to an
object that is keyed by all the algos we have.

This allows the miner calling the RPC to then choose the algo they want,
or even use one RPC call for mining in both algorithms.