Allow choosing algo for generate* RPCs.
Add a new, optional argument to generate and generatetoaddress that
allows to specify the PoW algo to be used for generating (defaults to
Neoscrypt).

For now, this just changes the actual algorithm that is used for
hashing; in the future (once merge-mining is implemented), selecting
'sha256d' will actually make the RPCs generate fake auxpows to get
properly merge-mined SHA256D blocks.

Also includes a new regtest for dual-algo; for now, it just verifies
that using the new argument chooses the algorithm correctly.  In the
future, we can use the same test to verify, for instance, that different
algorithms are accounted properly in the chain work.