Integrate PowAlgo in serialisation framework.
This "properly" integrates PowAlgo into Bitcoin's serialisation framework,
such that 'READWRITE(algo)' simply works without having to go through an
int manually.  This makes the code easier to read, but also avoids
duplicating that custom logic in the future when we need to store a
PowAlgo also as part of CDiskBlockIndex.