Handle dual-algo for GetBlockProofEquivalentTime.
With dual algo (that will be fully enabled in the future), we will have
one block *of each algorithm* per the target spacing time interval.  This
change prepares GetBlockProofEquivalentTime and its unit test for this,
by making sure to handle multiple algorithms correctly (even if they
have different difficulties and relative weights in the chain work).