Introduce relative weights of algos for chain work.
The two algorithms used for PoW (SHA256D and Neoscrypt) have an inherent
difference in the work they require per hash, which should be reflected
in the total chain work used to find the best chain.

This change introduces a relative weight factor (which was chosen to be
such that 2^10 SHA256D hashes equal one Neoscrypt hash, see discussion in
https://github.com/xyonplatform/xyon/issues/39).  This factor is used when
computing the work provided by each block.

We also add a regtest that verifies that a *shorter* chain of Neoscrypt
blocks can still be preferred to a *longer* chain of SHA256D blocks
(with the same difficulty for each algo).  We also had to adapt some tests
that specified a chain work explicitly, since we use Neoscrypt by default
for regtest chains and thus all chain work is scaled by a factor of 1024.
Tests that checked network hashes per second (which also changed now due
to the change in work for Neoscrypt) have been disabled for now, as we
will instead return the hashes per second *for each algo* later on and
then re-enable the tests after adapting them.