Fix mining algo already in CreateNewBlock.
In the future, block rewards will depend on the mining algo (per our
mining design).  This means that the PoW algo must already be known when
creating the block, so that the coinbase reward can be constructed
accordingly.  Thus unfortunately, it is not possible to let the miner
choose the algo only later.

This removes the capability in CBlockTemplate to keep two difficulties
and choosing the algo only later, and instead adds a new argument for
the chosen algo to CreateNewBlock.  It also adapts callers accordingly;
in particular, block caches for getblocktemplate and auxpow_miner now
take the algo into account when regenerating blocks, and the
auxpow_miner methods choose the appropriate mining algorithm already
(Neoscrypt for getwork and SHA256D for getauxblock).