Output PoW algo from getblocktemplate.
Add a new "algo" field in the result of getblocktemplate that indicates
which PoW algo was chosen for the block and should be used for mining it.
Also changes the code for getwork and getauxblock to return the algo
based on what is set on the block rather than a hardcoded string.