Make powLimit depend on mining algorithm.
The initial / minimum difficulty (powLimit in chain params) has to
reflect the different work required per hash for the mining algorithms.
This changes the params-configured value to be the limit for Neoscrypt
and introduces a function that computes the powLimit for each algo based
on that and the relative weight of each algo.

Since this makes calling CheckProofOfWork in the various places where
blocks are mined even more verbose, we add a new utility method for
checking the PoW of a PowData object and use that instead.