Remove pow's CheckProofOfWork.
After previous changes, the only remaining caller of CheckProofOfWork
from pow was PowData::checkProofOfWork (since all other places call that
function instead).

By inlining the logic from pow.cpp directly there, we can get rid of
CheckProofOfWork in pow itself.  This allows us to break the dependency
between powdata.cpp and pow.cpp, and move back pow.cpp to the "server"
library (where it is upstream).  This is good, as the current situation
was kind of broken anyway, since pow depends on chain, which is in
"server" and not "common".