Retarget difficulty for each algo independently.
This changes GetNextWorkRequired to only look at previous blocks of the
chosen algorithm, so that both algorithms retarget their difficulty
independently from each other.

To verify that the logic is correct, a new unit test checks that we get
exactly the same difficulty for each algo as we would get with the old
DGW function when we split the chain of blocks into separate chains for
each algorithm.