Update getdifficulty for dual algo.
This updates the getdifficulty and getmininginfo RPC methods to return the
current difficulty for each PoW algo.  It also removes the difficulty
field from getblockheader, getblock and getblockchaininfo.  Instead, the
difficulty value for a block can be found in the 'powdata' field now.

Since "chimaera-cli -getinfo" shows the difficulty and used to fetch it
from getblockchaininfo, this is now changed to fetch it (as an object)
from getmininginfo instead to keep the current behaviour.