Make CAuxPow internals actually private.
The internals (member variables) of CAuxPow are public, with a comment
that this is required for unit tests.  This commit makes them actually
private.  Legitimate uses are now covered through friend declarations
for the unit test and AuxpowToJSON and a new accessor for the parent
block (which is needed for PoW validation and mining).