Separate creation and initialisation of auxpow's.
Instead of both creating and already setting the "minimal" auxpow in
initAuxPow, this logic is split into two parts:  The construction of the
auxpow itself, and the logic that modifies the header.

Furthermore, we now return a reference to the parent block header from
initAuxPow, so it can be more easily mined.  This allows us to get rid
of the general accessor getParentBlockHeader, and replace it by just a
function returning the parent block hash (as before) to verify a PoW.