Do not inherit CMerkleTx from CAuxPow.
A CMerkleTx (for the parent block's coinbase tx) is a part of each
auxpow.  For that reason, CAuxPow was inheriting from CMerkleTx.  But
composition is more suitable here, as a CAuxPow "is not" a CMerkleTx.

This commit removes the inheritance and instead adds a CMerkleTx as
new (private) member variable to CAuxPow.