Add auxpow to PowData class.
This adapts the PowData class itself and its serialisation to support
merge-mined PoW's that hold a CAuxPow object instead of a fake header.

There is a unit test for basic serialisation / deserialisation in that
case, but otherwise this new support is not used anywhere yet.  In
particular, the consensus does not accept merge-mined blocks yet.