Re-enable upstream auxpow unit tests.
The auxpow-related unit tests (in auxpow_tests.cpp) have been commented
out while auxpow support was still missing in Xyon.  Now that we have it
back, we can re-enable the test for CAuxPow::check.

The other test is removed, since it is not relevant for Xyon:  The PoW
validation will be part of the powdata unit tests instead, and the
nVersion-related tests are obsolete anyway.