Allow merge-mined SHA256D blocks.
With this change to the validation logic of PowData, we allow (but not
yet require) SHA256D blocks to be merge mined.

Includes also a big refactoring of the powdata_tests to be split into
smaller pieces with custom test fixtures.  This allows us to build
cleaner unit tests and to properly share setup logic for the new tests
for validation of merge-mined PowData objects.