Make hash argument for getwork optional.
With dual-algo, we changed the interface for submitwork and getwork
to be "hash data" instead of just "data".  This makes sense, as it
leads to a more flexible, explicit interface and one that is consistent
with getauxblock.  However, legacy mining software won't be able to
make calls with this changed interface.

In this change, we make "hash" optional -- meaning that the old form
"getwork data" is supported as well to be backwards compatible with
existing mining software.  This is possible, since the data contains
the (mined) fake header, which in turn has the block hash in its
Merkle root field.