Split auxpow construction logic into two.
Instead of doing everything (constructing a fake auxpow, mining the
header and finalising the auxpow object) in computeAuxpow, we split the
logic into parts.

This allows us to reuse the construction / finalisation steps in the
future in a wrapper script that provides a getwork-like interface on top
of the core daemon's getauxblock.