Add script for solo-mining based on getauxblock.
This adds a small Python script that wraps around getauxblock in the
daemon and provides a getwork-like interface for mining.  The script
takes care of constructing a minimal auxpow (using the same code as
the regression tests).

While this is certainly not meant for mining in production, it can be
useful for testing and debugging purposes (and as a simple example for
how merge-mining works).