Add 'ismine' field to other name RPCs.
Similarly to 'name_pending', we can actually add the 'ismine' field
to all the RPCs that query for names.  When the user has a wallet
(and the daemon is actually compiled with wallet support), then the
address that is returned from the RPC already is checked against the
wallet.

This adds 'ismine' to name_show, name_history, name_scan and
name_filter.  It also adds an explicit test for the field in all
RPC methods (including name_pending and name_list, where it is
already now).

The field is *not* added to the REST interface for querying names.
This makes sense, as that interface may have lesser access restrictions
than the RPC interface and does not have the same existing wallet
infrastructure available.