Clean up of ownership RPC helper methods.
After adding the "ismine" field to all name RPC methods, we can clean
up the helper methods.  The "ismine" field can be added for all of the
help texts without requiring a separate function, and we can add a
utility method for getNameInfo with ownership as needed by most methods
in rpc/names.cpp.