Split out Namecoin-related parts of rpc/server.h.
rpc/server.h contains a block of definitions for functions related to
Namecoin RPC methods (and implemented in rpc/names.cpp).  In addition to
rpc/server.cpp, these are used also for wallet/rpcnames.cpp and rest.cpp.

This change is a trivial refactoring that just splits out these parts
into a new, Namecoin-specific header rpc/names.h.  rpc/server.h is
restored back to the exact version that is in upstream Bitcoin.  This
makes it easier to maintain in the future, and is also more explicit
about where this code is used.  It will simplify things if we extend the
set of shared functions for Namecoin-related RPCs in the future.