Add allowExisting option for name_new.
In https://github.com/namecoin/namecoin-core/issues/54, it was suggested
to disallow name_new for already existing names in the RPC interface by
default (while of course still allowing those on the consensus level).

This change leverages the new "options" argument for name_new to implement
this:  We add a new "allowExisting" option in the JSON object, false by
default, and raise an error for name_new with an existing name unless
this option is explicitly set to true.