Set premine P2SH address.
This hardcodes the final P2SH address that will hold the mainnet
premine: DHy2615XKevE23LVRVZVxGeqxadRGyiFW4

This address was created now by the founding members of the Xyon team.
It requires 2 of 4 keys to spend.  The team will distribute and burn
the coins from it after the presale is finished and the final
distribution is known.

This also fixes the testnet "timestamp" string and remines the genesis
blocks.  Regtest and testnet are now finished (barring discovery of any
issues not known so far), but the mainnet genesis block will be
recreated later just before the actual launch.

This fixes https://github.com/xyonplatform/xyon/issues/14.