Perform hashing inside buildNameNew.
The actual script of a NAME_NEW operation does not contain the name
or the rand value, but a hash combining both.  When the operation is
created through some interface, though, its inputs are typically the
name and rand itself.

This commit changes buildNameNew to not just build the name script out
of the hash value, but also perform the hashing.  All callers actually
have name and rand, so that this allows to share code between the RPC
methods name_new and namerawtransaction.  In the future, when name
operations are added to the namecoin-tx utility (see
https://github.com/namecoin/namecoin-core/issues/147), this function
can also be used there.