Support name operations in namecoin-tx.
This adds support for creating name operations through the namecoin-tx
utility.  This works very similar to namerawtransaction:  One has to
first create the transaction with the right inputs and outputs, and
can then apply one of the new commands to add the script prefix to the
name output.

See also https://github.com/namecoin/namecoin-core/issues/147.